Server-app/Company/Serializers.py
from django.db.models import fields
from rest_framework import serializers
from .models import *

class CompanySerializer(serializers.ModelSerializer):
    class Meta:
        model = Company
        fields = '__all__'

class JobSerializer(serializers.ModelSerializer):
    class Meta:
        model = Job
        fields = '__all__'    

class BlogSerializer(serializers.ModelSerializer):
    class Meta:
        model = Blog
        fields = '__all__'

class CatSerializer(serializers.ModelSerializer):
    class Meta:
        model = Category
        fields = '__all__'