Server-app/requirements.txt
APScheduler==3.7.0
asgiref==3.3.4
certifi==2020.12.5
chardet==4.0.0
Django==3.2.3
django-cors-headers==3.7.0
django-environ==0.4.5
django-optimized-image==0.3.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.6.0
idna==2.10
Pillow==8.2.0
psycopg2==2.8.6
PyJWT==2.1.0
pytz==2021.1
requests==2.25.1
six==1.16.0
sqlparse==0.4.1
tinify==1.5.1
tzlocal==2.1
urllib3==1.26.4
